+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project will adhere to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+### Added
+
+- Vector.toString method
+
 ## [0.0.11] - 2025-03-16
 
 ### Added